requirements.txt
# Deep Learning & ML
tensorflow>=2.10.0
keras>=2.10.0
scikit-learn>=1.0.0

# Computer Vision
opencv-python>=4.7.0
mediapipe>=0.10.0
pillow>=9.0.0

# Data Processing
numpy>=1.23.0
pandas>=1.5.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Web Interface
streamlit>=1.20.0

# Text-to-Speech
pyttsx3>=2.90
gTTS>=2.3.0

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
python-dotenv>=0.21.0

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.20.0

# Model Export & Optimization
onnx>=1.13.0
onnxruntime>=1.14.0

# Kaggle API (for dataset download)
kaggle>=1.5.0
